reworked structure + added GUImain.py for GUI
