+++ main.py
@@ -1,30 +1,28 @@
-from modules import user_input
-from modules import create_tables, migrate_csv_to_db
-from modules.extract_gz_to_csv import extract_gz_to_csv
-from modules.Migrate_csv_to_db import migrate_csv_to_db
-from modules.db_data_requests.data_user_input import display_data_func
-from modules import delete_db_data,delete_files
+from modules.DOWNLOAD_DATA import user_input,extract_gz_to_csv,Migrate_csv_to_db
+from modules.DATABASE import create_db,delete_db_data,delete_files
+from modules.CONSOLE import data_user_input
+from modules.GUI import FeinstaubGUI
+
 
 def main():
 
     #Datenbank erstellen
-    create_tables("feinstaub.db")
-
-    user_input.inputDate()
+    create_db.create_tables("feinstaub.db")
 
 
-    #Convertierung der csv.gz dateiten in csv dateien
-    extract_gz_to_csv()
+    #Abfrage für Download
+    user_input.input_for_download()
+    #download + entpacken
+    extract_gz_to_csv.extract_gz_to_csv()
+    Migrate_csv_to_db.migrate_csv_to_db()
 
-    #Starte den Import der CSV-Dateien in die Datenbank
-    migrate_csv_to_db()
+    #Abfrage der Daten für Zeitraum X von Min/Max/Avg
+    data_user_input.display_data_func()
 
-   #Zeige gewünschte Werte an
-    display_data_func()
-
-    #lösche CSV und Datenbank:
+    #Daten löschen
     delete_files.delete_GZ_and_csv()
     delete_db_data.delete_all_data()
+
 
 if __name__ == "__main__":
     main()
