--- modules/__init__.py
@@ -1,3 +1,0 @@
-from .create_db import create_tables
-from .Migrate_csv_to_db import migrate_csv_to_db
-from .extract_gz_to_csv import extract_gz_to_csv
